+++ resources/modules/editor/src/js/classes/elements/BaseElement.js
@@ -41,6 +41,24 @@ class BaseElement extends ControlStack {
   }
 
   /**
+   *
+   * @returns {BaseElement|null}
+   */
+  findElementById(elementId){
+    if(! elementId){
+      return null
+    }
+    if(this.getId() === elementId){
+      return this
+    }
+    for(const el of this.getChildren()){
+      if(el.findElementById(elementId)){
+        return el.findElementById(elementId)
+      }
+    }
+    return null
+  }
+  /**
    * Задать настройки
    * @param settings
    * @param settingsLock
@@ -196,7 +214,7 @@ class BaseElement extends ControlStack {
   }
 
   getChildren() {
-    if (!this.children.length) {
+    if (!this.children?.length) {
       return [];
     }
     return this.children;