+++ resources/modules/editor/src/js/store/current-context-element/actions.js
@@ -1,9 +1,10 @@
 export const SET_CURRENT_CONTEXT_ELEMENT = 'SET_CURRENT_CONTEXT_ELEMENT';
 
-export function setCurrentContextElement(element) {
+export function setCurrentContextElement(element, inMainWindow =false) {
   return {
     type: SET_CURRENT_CONTEXT_ELEMENT,
-    element
+    element,
+    inMainWindow
   };
 }